+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Zero-width non-joiners are now stripped to prevent output issues, similar to non-breaking whitespace
+
 ### Fixed
 - Fix namespace in composer [#67](https://github.com/soundasleep/html2text/pull/67)